+++ DataManager/ClientApp/minerva/ts/views/reports/dashboard.view.tsx
@@ -1,17 +1,18 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { DashboardLayout, ReportDefinition } from 'minerva/core';
+import { ProductDashboard, ReportDefinition, UpdateProductDashboardModel } from 'minerva/core';
 import { ProductLayout, ProductSelector } from 'minerva/components';
 import { ReportComponent } from './report.component';
 
 export function DashboardView() {
   const params = useParams(),
         [productId, setProductId] = useState(''),
+        [dashboardId, setDashboardId] = useState(''),
         [dashboardName, setDashboardName] = useState(''),
         [layout, setLayout] = useState<ReportDefinition[][]>([]);
 
   useEffect(() => {
-    const layout$ = DashboardLayout.subscribe(value => {
+    const layout$ = ProductDashboard.subscribe(value => {
       setDashboardName(value.name);
       const layout = value.reports.groupBy('rowIndex').orderBy('key')
           .map((row: { key: string, elements: ReportDefinition[] }) => {
@@ -25,8 +26,17 @@ export function DashboardView() {
   useEffect(() => {
     const { productId, dashboardId } = params;
     setProductId(productId);
-    DashboardLayout.loadWithSplashScreen({ pathParams: { productId, dashboardId } });
+    setDashboardId(dashboardId);
+    ProductDashboard.loadWithSplashScreen({ pathParams: { productId, dashboardId } });
   }, [params]);
+
+  async function saveDashboardReports() {
+    const dashboard = await UpdateProductDashboardModel.fetch({
+      pathParams: { productId, dashboardId },
+      body: { productId, dashboardId, reports: layout.flatMap(x => x) },
+    });
+    ProductDashboard.loadData(dashboard);
+  }
 
   return <ProductLayout>
     <ProductSelector navPath='/dashboard'>
@@ -36,7 +46,7 @@ export function DashboardView() {
           <span className="fa fa-ellipsis" />
         </button>
         <div className="dropdown-menu">
-          <button className="dropdown-item">Save Dashboard</button>
+          <button className="dropdown-item" onClick={() => saveDashboardReports()}>Save Dashboard</button>
         </div>
       </div>
     </ProductSelector>